core/color_processor.py
from PIL import Image
import numpy as np
from sklearn.cluster import KMeans

class ColorProcessor:
    def __init__(self, image):
        self.image = image
        self.color_ranges = []
        self.max_colors = 120

    def process_colors(self):
        # Step 1: Extract colors
        self.extract_colors()
        # Step 2: Limit colors
        self.limit_colors()
        # Step 3: Group colors
        self.group_colors()
        # Step 4: Replace excess colors
        self.replace_excess_colors()

    def extract_colors(self):
        self.image_data = np.array(self.image)
        self.pixels = self.image_data.reshape(-1, 3)

    def limit_colors(self):
        # Use KMeans clustering to reduce colors
        kmeans = KMeans(n_clusters=self.max_colors)
        kmeans.fit(self.pixels)
        self.limited_colors = kmeans.cluster_centers_.astype(int)
        self.labels = kmeans.labels_

    def group_colors(self):
        # Group colors into ranges of 10 to 12, maximum of 10 ranges
        colors = self.limited_colors.tolist()
        self.color_ranges = [colors[i:i+12] for i in range(0, len(colors), 12)]
        self.color_ranges = self.color_ranges[:10]

    def replace_excess_colors(self):
        # Map original pixels to limited colors
        # Create a mapping from labels to colors in the color ranges
        # Replace colors in the image data
        pass  # Implement mapping and replacement

    def update_image(self):
        # Update the image with the new colors
        new_pixels = self.limited_colors[self.labels]
        new_image_data = new_pixels.reshape(self.image_data.shape)
        self.image = Image.fromarray(new_image_data.astype('uint8'), 'RGB')
