requirements.txt
Pillow
numpy
scikit-learn
reportlab
